refactor: ktor file organization
