+++ ktor/src/main/kotlin/com/chesspuzzletext2sql/features/puzzles/domains/LLMConfig.kt
@@ -1,4 +1,4 @@
-package com.chesspuzzletext2sql.features.puzzleSearch.models
+package com.chesspuzzletext2sql.features.puzzles.domains
 
 data class LLMConfig(
     val provider: String,
