+++ ktor/src/main/kotlin/com/chesspuzzletext2sql/shared/config/ApplicationConfig.kt
@@ -1,6 +1,6 @@
-package com.chesspuzzletext2sql.config
+package com.chesspuzzletext2sql.shared.config
 
-import com.chesspuzzletext2sql.errors.StartupError
+import com.chesspuzzletext2sql.shared.errors.StartupError
 import com.github.michaelbull.result.Result
 import com.github.michaelbull.result.mapError
 import com.github.michaelbull.result.zipOrAccumulate
