+++ ktor/src/main/kotlin/com/chesspuzzletext2sql/shared/data/Schemas.kt
@@ -1,9 +1,8 @@
-package com.chesspuzzletext2sql.features.puzzleSearch.data
+package com.chesspuzzletext2sql.shared.data
 
 import org.jetbrains.exposed.v1.core.Table
 
-object PuzzleTable : Table("T_Puzzle") {
-    val id = integer("id").autoIncrement()
+object Puzzles : Table("T_Puzzle") {
     val puzzleId = varchar("puzzle_id", 255)
     val fen = varchar("fen", 255)
     val moves = varchar("moves", 255)
@@ -13,7 +12,7 @@ object PuzzleTable : Table("T_Puzzle") {
     val nbPlays = integer("nb_plays")
     val themes = varchar("themes", 255)
     val gameUrl = varchar("game_url", 255)
-    val openingTags = varchar("opening_tags", 255)
+    val openingTags = varchar("opening_tags", 255).nullable()
 
-    override val primaryKey = PrimaryKey(id)
+    override val primaryKey = PrimaryKey(puzzleId)
 }
